refactor: Refactor import statements and directory structure
The import statements for components have been updated to match a change in directory structure. Component files have been moved up one level from a redundant subfolder within their respective component folders. This reorganization simplifies the file structure and improves the readability and maintainability of import statements in the codebase.
