+++ src/components/interface/triangle.tsx
