+++ src/root.tsx
@@ -1,7 +1,7 @@
 import { component$ } from "@builder.io/qwik";
 import { QwikCityProvider, RouterOutlet, ServiceWorkerRegister } from "@builder.io/qwik-city";
 
-import { RouterHead } from "~/components/function/router-head/router-head";
+import { RouterHead } from "~/components/function/router-head";
 
 import "./global.css";
