+++ src/routes/hello/components/game.tsx
@@ -1,6 +1,6 @@
 import { component$ } from "@builder.io/qwik";
 
-import { Button } from "~/components/interface/button/button";
+import { Button } from "~/components/interface/button";
 
 import { Food } from "./food";
 import styles from "./game.module.css";
