+++ src/routes/hello/components/key.tsx
@@ -1,6 +1,6 @@
 import { component$ } from "@builder.io/qwik";
 
-import { Triangle, type TriangleProps } from "~/components/interface/triangle/triangle";
+import { Triangle, type TriangleProps } from "~/components/interface/triangle";
 
 import styles from "./key.module.css";
