+++ src/routes/layout.tsx
@@ -1,6 +1,6 @@
 import { component$, Slot } from "@builder.io/qwik";
 
-import { Footer } from "~/components/interface/footer/footer";
+import { Footer } from "~/components/interface/footer";
 import { Header } from "~/components/interface/header/header";
 import { useGlobalStyles } from "~/hooks/global-styles/global-styles";
